test: extract db access guard test helpers
